Remove old analysis script / rename current one
